Extract `BuildFileSettings` to own file and convert tests to SwiftTesting
